requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.124.4
frozenlist==1.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
json-logging==1.5.1
motor==3.7.1
multidict==6.7.0
passlib==1.7.4
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
PyYAML==6.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
